requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=4.7.0=pyhd8ed1ab_0
appnope=0.1.4=pyhd8ed1ab_1
argon2-cffi=23.1.0=pyhd8ed1ab_1
argon2-cffi-bindings=21.2.0=py313h20a7fcf_5
arrow=1.3.0=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
async-lru=2.0.4=pyhd8ed1ab_1
attrs=24.3.0=pyh71513ae_0
babel=2.16.0=pyhd8ed1ab_1
beautifulsoup4=4.12.3=pyha770c72_1
bleach=6.2.0=pyhd8ed1ab_2
brotli=1.1.0=hd74edd7_2
brotli-bin=1.1.0=hd74edd7_2
brotli-python=1.1.0=py313h3579c5c_2
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.12.14=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2024.12.14=pyhd8ed1ab_0
cffi=1.17.1=py313hc845a76_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=py313h0ebd0e5_0
coverage=7.6.10=py313ha9b7d5b_0
cycler=0.12.1=pyhd8ed1ab_1
debugpy=1.8.11=py313h928ef07_0
decorator=5.1.1=pyhd8ed1ab_1
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_1
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
fonttools=4.55.3=py313ha9b7d5b_1
fqdn=1.5.1=pyhd8ed1ab_1
freetype=2.12.1=hadb7bae_2
h11=0.14.0=pyhd8ed1ab_1
h2=4.1.0=pyhd8ed1ab_1
hpack=4.0.0=pyhd8ed1ab_1
httpcore=1.0.7=pyh29332c3_1
httpx=0.28.1=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_1
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.5.0=pyha770c72_1
importlib_resources=6.4.5=pyhd8ed1ab_1
iniconfig=2.0.0=pyhd8ed1ab_1
ipykernel=6.29.5=pyh57ce528_0
ipython=8.31.0=pyh707e725_0
ipywidgets=8.1.5=pyhd8ed1ab_1
isoduration=20.11.0=pyhd8ed1ab_1
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.5=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_1
json5=0.10.0=pyhd8ed1ab_1
jsonpointer=3.0.0=py313h8f79df9_1
jsonschema=4.23.0=pyhd8ed1ab_1
jsonschema-specifications=2024.10.1=pyhd8ed1ab_1
jsonschema-with-format-nongpl=4.23.0=hd8ed1ab_1
jupyter=1.1.1=pyhd8ed1ab_1
jupyter-lsp=2.2.5=pyhd8ed1ab_1
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_console=6.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
jupyter_events=0.11.0=pyhd8ed1ab_0
jupyter_server=2.15.0=pyhd8ed1ab_0
jupyter_server_terminals=0.5.3=pyhd8ed1ab_1
jupyterlab=4.3.4=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_2
jupyterlab_server=2.27.3=pyhd8ed1ab_1
jupyterlab_widgets=3.0.13=pyhd8ed1ab_1
kiwisolver=1.4.7=py313hf9c7212_0
krb5=1.21.3=h237132a_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=26_osxarm64_openblas
libbrotlicommon=1.1.0=hd74edd7_2
libbrotlidec=1.1.0=hd74edd7_2
libbrotlienc=1.1.0=hd74edd7_2
libcblas=3.9.0=26_osxarm64_openblas
libcxx=19.1.6=ha82da77_1
libdeflate=1.23=hec38601_0
libedit=3.1.20191231=hc8eb9b7_2
libexpat=2.6.4=h286801f_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=26_osxarm64_openblas
liblzma=5.6.3=h39f12f2_1
libmpdec=4.0.0=h99b78c6_0
libopenblas=0.3.28=openmp_hf332438_1
libpng=1.6.44=hc14010f_0
libsodium=1.0.20=h99b78c6_0
libsqlite=3.47.2=h3f77e49_0
libtiff=4.7.0=h551f018_3
libwebp-base=1.5.0=h2471fea_0
libxcb=1.17.0=hdb1d25a_0
libzlib=1.3.1=h8359307_2
llvm-openmp=19.1.6=hdb05f8b_0
markupsafe=3.0.2=py313ha9b7d5b_1
matplotlib=3.10.0=py313h39782a4_0
matplotlib-base=3.10.0=py313haaf02c0_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mistune=3.1.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
nb_conda_kernels=2.5.1=pyh707e725_2
nbclient=0.10.2=pyhd8ed1ab_0
nbconvert-core=7.16.5=pyhd8ed1ab_0
nbformat=5.10.4=pyhd8ed1ab_1
ncurses=6.5=h7bae524_1
nest-asyncio=1.6.0=pyhd8ed1ab_1
notebook=7.3.2=pyhd8ed1ab_0
notebook-shim=0.2.4=pyhd8ed1ab_1
numpy=2.2.1=py313ha4a2180_0
openjpeg=2.5.3=h8a3d83b_0
openssl=3.4.0=h39f12f2_0
overrides=7.7.0=pyhd8ed1ab_1
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py313h47b39a6_1
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_1
patsy=1.0.1=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=py313hb37fac4_0
pip=24.3.1=pyh145f28c_2
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_2
platformdirs=4.3.6=pyhd8ed1ab_1
pluggy=1.5.0=pyhd8ed1ab_1
prometheus_client=0.21.1=pyhd8ed1ab_0
prompt-toolkit=3.0.48=pyha770c72_1
prompt_toolkit=3.0.48=hd8ed1ab_1
psutil=6.1.1=py313h90d716c_0
pthread-stubs=0.4=hd74edd7_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pycparser=2.22=pyh29332c3_1
pygments=2.18.0=pyhd8ed1ab_1
pyobjc-core=10.3.2=py313hb6afeec_0
pyobjc-framework-cocoa=10.3.2=py313hb6afeec_0
pyparsing=3.2.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha55dd90_7
pytest=8.3.4=pyhd8ed1ab_1
pytest-cov=6.0.0=pyhd8ed1ab_1
python=3.13.1=h4f43103_102_cp313
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pyhd8ed1ab_1
python-fastjsonschema=2.21.1=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2024.2=pyhd8ed1ab_1
python_abi=3.13=5_cp313
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.2=py313h20a7fcf_1
pyzmq=26.2.0=py313h0e8b002_3
qhull=2020.2=h420ef59_5
readline=8.2=h92ec313_1
referencing=0.35.1=pyhd8ed1ab_1
requests=2.32.3=pyhd8ed1ab_1
rfc3339-validator=0.1.4=pyhd8ed1ab_1
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.22.3=py313hdde674f_0
scikit-learn=1.6.0=py313hecba28c_0
scipy=1.14.1=py313hc010ede_2
seaborn=0.13.2=hd8ed1ab_3
seaborn-base=0.13.2=pyhd8ed1ab_3
send2trash=1.8.3=pyh31c8845_1
setuptools=75.6.0=pyhff2d567_1
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pyhd8ed1ab_1
soupsieve=2.5=pyhd8ed1ab_1
stack_data=0.6.3=pyhd8ed1ab_1
statsmodels=0.14.4=py313h93df234_0
terminado=0.18.1=pyh31c8845_0
threadpoolctl=3.5.0=pyhc1e730c_0
tinycss2=1.4.0=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
toml=0.10.2=pyhd8ed1ab_1
tomli=2.2.1=pyhd8ed1ab_1
tornado=6.4.2=py313h90d716c_0
traitlets=5.14.3=pyhd8ed1ab_1
types-python-dateutil=2.9.0.20241206=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
typing_utils=0.1.0=pyhd8ed1ab_1
tzdata=2024b=hc8b5060_0
uri-template=1.3.0=pyhd8ed1ab_1
urllib3=2.3.0=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_1
webcolors=24.11.1=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_3
websocket-client=1.8.0=pyhd8ed1ab_1
widgetsnbextension=4.0.13=pyhd8ed1ab_1
xorg-libxau=1.0.12=h5505292_0
xorg-libxdmcp=1.1.5=hd74edd7_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_1
zstandard=0.23.0=py313hf2da073_1
zstd=1.5.6=hb46c0d2_0
